+++ src/main/java/org/hankki/hankkiserver/api/favorite/service/FavoriteCommandService.java
@@ -1,10 +1,12 @@
-package org.hankki.hankkiserver.api.service.favorite;
+package org.hankki.hankkiserver.api.favorite.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.hankki.hankkiserver.api.favorite.controller.request.FavoritePostRequest;
+import org.hankki.hankkiserver.api.favorite.service.FavoriteAdpater;
 import org.hankki.hankkiserver.domain.favorite.model.Favorite;
 import org.hankki.hankkiserver.domain.favorite.repository.FavoriteRepository;
-import org.hankki.hankkiserver.api.service.favorite.command.FavoritePostCommand;
+import org.hankki.hankkiserver.api.favorite.service.command.FavoritePostCommand;
 import org.hankki.hankkiserver.domain.user.model.User;
 import org.hankki.hankkiserver.domain.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,7 @@ public class FavoriteCommandService {
 
   private final FavoriteRepository favoriteRepository;
   private final UserRepository userRepository;
+  private final FavoriteAdpater favoriteAdpater;
 
   @Transactional
   public Long create(final FavoritePostCommand favoritePostCommand) {
@@ -28,6 +31,14 @@ public class FavoriteCommandService {
     favoriteRepository.save(favorite);
 
     return favorite.getId();
+
+  }
+
+  @Transactional
+  public void delete(final long favoriteId) {
+
+    Favorite favorite = favoriteAdpater.findById(favoriteId);
+    favoriteRepository.delete(favorite);
 
   }
 }
